+++ src/features/lost/components/upload/ImageSection.types.ts
@@ -1,0 +1,4 @@
+export type ImageSectionProps = {
+  image: File | null;
+  setImage: (file: File | null) => void;
+};
