+++ src/features/lost/components/upload/TimeSelect.types.ts
@@ -1,0 +1,8 @@
+export type TimeSelectProps = {
+  selectedDay: string;
+  setSelectedDay: (day: string) => void;
+  time: string;
+  setTime: (time: string) => void;
+  selectOpen: boolean;
+  setSelectOpen: (open: boolean) => void;
+};
